+++ flexvalue/calculations.py
@@ -549,6 +549,14 @@ class FlexValueRun:
             .reset_index(drop=True)
         )
 
+    def get_all_trc_gas_benefits_df(self, user_inputs):
+        return (
+            pd.concat([                
+                flx_project.calculate_trc_gas_benefits()
+                for flx_project in self.get_flexvalue_projects(user_inputs).values()
+            ])
+        )
+
     def get_total_trc_gas_benefits(self, user_inputs):
         """The total gas benefits across all FlexValueProjects
 
@@ -681,10 +689,14 @@ class FlexValueRun:
             user_inputs=user_inputs
         )
         elec_benefits = self.get_all_trc_electricity_benefits_df(user_inputs)
-        gas_benefits = self.get_total_trc_gas_benefits(user_inputs)
-
+        gas_benefits = self.get_total_trc_gas_benefits(user_inputs)      
         # if index wasn't already set with the ID colum, set it for joining to the output
         if user_inputs.index.name != "ID":
             user_inputs = user_inputs.set_index("ID")
         outputs_table = user_inputs.join(outputs_table).reset_index()
         return outputs_table, outputs_table_totals, elec_benefits, gas_benefits
+
+    def get_time_series_results(self, user_inputs):
+        elec_benefits = self.get_all_trc_electricity_benefits_df(user_inputs)
+        gas_benefits = self.get_all_trc_gas_benefits_df(user_inputs)      
+        return elec_benefits, gas_benefits